% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakepkg.R
\docType{package}
\name{fakepkg}
\alias{fakepkg}
\title{An R package for reproduciable furrr problem}
\seealso{
Useful links:

\url{https://github.com/jinyancool/fakepkg}

Report bugs at \url{https://github.com/jinyancool/fakepkg/issues}
}
\author{
Jinyan HUANG
}
